requirements.txt
openai>=2.0.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
